requirements.txt
django > 1.8, <= 1.10
djangorestframework >= 3.0, < 4.0
